requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3